+++ src/components/Chooser/Chooser.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import { array, func, instanceOf } from 'prop-types';
+import { array, func, number, instanceOf } from 'prop-types';
 import SelectButton from '../SelectButton';
 import Collection from '../../core/Collection';
 
-const Chooser = ({ frameworks, visible, onFrameworkSelect }) => {
+const Chooser = ({ frameworks, minimum, maximum, visible, onFrameworkSelect }) => {
   return (
     <div className="section has-text-centered">
-      <div className="container">
+      <div className="container is-fullhd">
         <h2 className="subtitle is-5">
-          Select <strong>at least 2</strong> and <strong>maximum 4</strong> front-end frameworks and
-          compare them to each other.
+          Select <strong>at least {minimum}</strong> and <strong>maximum {maximum}</strong>{' '}
+          front-end frameworks and compare them to each other.
         </h2>
         <div className="notification">
           {frameworks.map(framework => {
@@ -34,7 +34,14 @@ const Chooser = ({ frameworks, visible, onFrameworkSelect }) => {
 Chooser.propTypes = {
   frameworks: array.isRequired,
   visible: instanceOf(Collection).isRequired,
-  onFrameworkSelect: func.isRequired
+  onFrameworkSelect: func.isRequired,
+  minimum: number,
+  maximum: number
+};
+
+Chooser.defaultProps = {
+  minimum: 2,
+  maximum: 4
 };
 
 export default Chooser;